pyproject.toml
[project]
name = "doom-vlm"
version = "0.1.0"
description = "VLM agents play DOOM via VizDoom"
requires-python = ">=3.11"
dependencies = [
    "vizdoom>=1.2.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
]

[project.scripts]
doom-vlm = "doom_vlm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
